Consolidate debug utility functions to a namespace
Created the debug namespace to replace the debugManager class and the debug components file.  Circumvented circular dependency in the application class and debug function by removing the class' builtin logging function and making it use the debug  namespace instead.  Also, variables relating to debug logging and application management have been wrapped in a nameless namespace to ensure they are scoped to the debug source and header files instead of being global.